feat: 移除旧版本datepicker 样式
