+++ style/web/components/_docs.less
@@ -38,8 +38,6 @@
 
 @import "./tooltip/_docs.less";
 
-@import "./date-picker/_docs.less";
-
 @import "./swiper/_docs.less";
 
 @import "./form/_docs.less";
