--- style/web/components/date-picker-new/_index.less
@@ -1,359 +1,0 @@
-// 组件允许单个组件打包，因此默认引入公共基础样式
-
-@import "../../base.less";
-
-@import "./_var.less";
-
-@import "./_mixin.less";
-
-.@{prefix}-date-picker,
-.@{prefix}-date-range-picker {
-  display: inline-flex;
-}
-
-.@{prefix}-date-picker__panel-container,
-.@{prefix}-date-range-picker__panel-container {
-  padding: 0;
-  min-width: @datepicker-panel-container-min-width;
-}
-
-.@{prefix}-date-picker__panel,
-.@{prefix}-date-range-picker__panel {
-  display: flex;
-  flex-direction: column;
-
-  &--direction-row {
-    flex-direction: row;
-  }
-
-  .@{prefix}-select__dropdown {
-    max-height: @datepicker-panel-select-max-width;
-  }
-
-  // 配合time-picker使用 需重置部分样式
-  .@{prefix}-time-picker__panel {
-    width: @datepicker-time-panel-width;
-
-    &-section-body {
-      padding: 0;
-    }
-
-    &-body {
-      margin: 0;
-    }
-  }
-}
-
-.@{prefix}-date-picker__panel--content,
-.@{prefix}-date-range-picker__panel--content,
-.@{prefix}-date-range-picker__panel--content-wrapper {
-  display: flex;
-}
-.@{prefix}-date-range-picker__panel--content-wrapper {
-  gap: 6px;
-}
-
-.@{prefix}-date-picker__header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: @spacer-l;
-  width: 100%;
-
-  &-controller {
-    display: inline-flex;
-    gap: @spacer;
-
-    .@{prefix}-date-picker__header-controller--month {
-      width: @datepicker-month-controller-width;
-    }
-
-    .@{prefix}-date-picker__header-controller--year {
-      width: @datepicker-year-controller-width;
-    }
-  }
-}
-
-.@{prefix}-date-picker__table {
-  table {
-    width: 100%;
-    border-collapse: collapse;
-  }
-
-  th {
-    text-align: center;
-    color: @datepicker-th-font-color;
-    font-weight: @datepicker-th-font-weight;
-  }
-
-  td {
-    text-align: center;
-    font-weight: @datepicker-font-weight-medium;
-  }
-
-  thead::after {
-    content: "-";
-    display: block;
-    line-height: @spacer-s - 1px;
-    color: transparent;
-  }
-}
-
-.@{prefix}-date-picker__footer {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  padding: @datepicker-presets-padding;
-  gap: @datepicker-presets-gap;
-
-  &--bottom {
-    border-top: 1px solid @datepicker-panel-divider-color;
-  }
-
-  &--top {
-    border-bottom: 1px solid @datepicker-panel-divider-color;
-  }
-
-  &--left {
-    flex-direction: column;
-    border-right: 1px solid @datepicker-panel-divider-color;
-  }
-
-  &--right {
-    flex-direction: column;
-    border-left: 1px solid @datepicker-panel-divider-color;
-  }
-}
-
-.@{prefix}-date-picker__presets {
-  display: flex;
-  gap: @datepicker-presets-gap;
-  white-space: nowrap;
-  min-width: @datepicker-presets-min-width;
-
-  .@{prefix}-button {
-    cursor: pointer;
-  }
-
-  a {
-    color: @datepicker-presets-text-color;
-
-    &:hover {
-      color: @datepicker-presets-text-color-hover;
-    }
-  }
-}
-
-.@{prefix}-date-picker__panel--time,
-.@{prefix}-date-range-picker__panel--time {
-  display: flex;
-  flex-direction: column;
-  gap: 2px;
-  padding: @datepicker-time-panel-padding;
-  border-left: 1px solid @datepicker-panel-divider-color;
-
-  &-viewer {
-    display: flex;
-    height: 32px;
-    line-height: 22px;
-    align-items: center;
-    justify-content: center;
-    color: @text-color-primary;
-  }
-
-  .@{prefix}-time-picker__panel-body {
-    margin: 0;
-  }
-}
-
-.@{prefix}-date-picker__panel--year,
-.@{prefix}-date-picker__panel--month,
-.@{prefix}-date-picker__panel--date {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  padding: @spacer-l;
-  gap: @spacer-l;
-  max-width: @datepicker-panel-max-width;
-}
-
-.@{prefix}-date-picker__panel--year,
-.@{prefix}-date-picker__panel--month {
-  .@{prefix}-date-picker__header-controller--year {
-    width: 130px;
-  }
-
-  .@{prefix}-date-picker__table {
-    table {
-      tbody {
-        width: 100%;
-        display: flex;
-        gap: @spacer-2;
-        flex-direction: column;
-
-        tr {
-          display: flex;
-          justify-content: space-between;
-        }
-      }
-    }
-  }
-
-  .@{prefix}-date-picker__cell--disabled + .@{prefix}-date-picker__cell--disabled:before {
-    left: -@spacer-7;
-  }
-
-  .@{prefix}-date-picker__cell:nth-child(1) {
-    text-align: left;
-  }
-
-  .@{prefix}-date-picker__cell:nth-child(3) {
-    text-align: right;
-  }
-
-  .@{prefix}-date-picker__cell-inner {
-    width: @datepicker-cell-min-width-l;
-  }
-}
-
-.@{prefix}-date-picker__cell {
-  cursor: pointer;
-  position: relative;
-  padding: 0;
-
-  &::before,
-  &::after {
-    content: "";
-    position: absolute;
-    top: 50%;
-    right: 0;
-    left: 0;
-    z-index: 1;
-    height: @datepicker-cell-height;
-    transform: translateY(-50%);
-    transition: all .2s @anim-time-fn-easing;
-  }
-
-  &:hover {
-    .@{prefix}-date-picker__cell-inner {
-      box-shadow: inset 0 0 0 1px @brand-color;
-    }
-  }
-
-  &-inner {
-    position: relative;
-    z-index: 5;
-    display: inline-flex;
-    justify-content: center;
-    align-items: center;
-    width: @datepicker-cell-min-width;
-    height: @datepicker-cell-height;
-    margin: @spacer-s - 1px;
-    border-radius: @datepicker-panel-border-radius;
-    transition: @datepicker-transition;
-  }
-
-  &--now {
-    .@{prefix}-date-picker__cell-inner {
-      color: @datepicker-now-text-color;
-      background: @datepicker-now-background-color;
-    }
-  }
-
-  &--hover-highlight {
-    &::after {
-      background-color: @datepicker-highlight-color;
-    }
-  }
-
-  &--hover-highlight + &--hover-highlight {
-    &::after {
-      left: -28px;
-    }
-  }
-
-  &--highlight + &--highlight {
-    &::before {
-      left: -28px;
-    }
-  }
-
-  &--highlight {
-    &::before {
-      background-color: @datepicker-highlight-color;
-    }
-
-    &.@{prefix}-date-picker__cell--hover-highlight {
-      &::after {
-        background-color: @datepicker-hover-highlight-color;
-      }
-    }
-  }
-
-  &--additional {
-    &:hover {
-      .@{prefix}-date-picker__cell-inner {
-        box-shadow: none;
-        border-color: @datepicker-cell-hover-background;
-        background-color: @datepicker-cell-hover-background;
-      }
-    }
-
-    .@{prefix}-date-picker__cell-inner {
-      color: @datepicker-cell-disabled-color;
-    }
-  }
-
-  &--active {
-    .@{prefix}-date-picker__cell-inner {
-      color: @datepicker-cell-active-color;
-      background-color: @datepicker-cell-active-background;
-    }
-
-    &-start {
-      &::before {
-        left: 6px;
-        border-top-left-radius: @border-radius;
-        border-bottom-left-radius: @border-radius;
-      }
-    }
-
-    &-end {
-      &::before {
-        right: 6px;
-        border-top-right-radius: @border-radius;
-        border-bottom-right-radius: @border-radius;
-      }
-    }
-  }
-
-  &--disabled {
-    &:hover {
-      .@{prefix}-date-picker__cell-inner {
-        box-shadow: none;
-        background-color: @datepicker-cell-disable-background-color;
-      }
-    }
-
-    .@{prefix}-date-picker__cell-inner {
-      cursor: not-allowed;
-      color: @datepicker-cell-disabled-color;
-      background-color: @datepicker-cell-disable-background-color;
-    }
-  }
-
-  &--disabled + &--disabled {
-    &::before {
-      left: -@spacer-4;
-      right: @spacer-l;
-      background-color: @datepicker-cell-disable-background-color;
-    }
-  }
-}
-
-// hover 日期淡化输入框
-.@{prefix}-date-picker__input--placeholder {
-  input {
-    color: @text-color-placeholder;
-  }
-}
