--- style/web/components/date-picker-new/_var.less
@@ -1,56 +1,0 @@
-// 组件变量
-// 名称可按如下规则定义：
-// <component>-[type]-[attrtype]-<attr>-[status]
-
-// component:组件名，如button，
-// type: 组件类型，如 button 的次要按钮（line）
-// attrtype: 属性的具体应用场景。如颜色，用于背景（bg）、文本（text）、边框（border）等
-// attr: 属性名称，如color、height、radius等
-// status: 表示组件状态或尺寸，如 hover、disabled、s、l 等
-
-// 如：@button-line-bg-color-hover
-// 如：@button-line-height-s
-
-@datepicker-th-font-color: @text-color-secondary;
-
-@datepicker-panel-max-width: 280px;
-@datepicker-panel-container-min-width: 280px;
-@datepicker-panel-select-max-width: 160px;
-@datepicker-panel-divider-color: @gray-color-3;
-@datepicker-time-panel-width: 216px;
-@datepicker-time-panel-padding: @spacer-l;
-
-@datepicker-year-controller-width: 78px;
-@datepicker-month-controller-width: 80px;
-
-@datepicker-cell-min-width: 24px;
-@datepicker-cell-height: 24px;
-@datepicker-cell-min-width-l: 48px;
-
-@datepicker-th-font-weight: 400;
-@datepicker-font-weight-medium: 500;
-@datepicker-panel-border-radius: @border-radius;
-
-@datepicker-presets-padding: @spacer-l;
-@datepicker-presets-gap: @spacer;
-
-// cell
-@datepicker-cell-hover-background: @bg-color-container-hover;
-@datepicker-cell-active-color: @text-color-anti;
-@datepicker-cell-active-background: @brand-color;
-
-@datepicker-cell-disabled-color: @text-color-disabled;
-@datepicker-cell-disable-border-color: @gray-color-4;
-@datepicker-cell-disable-background-color: @gray-color-2;
-
-@datepicker-now-text-color: @brand-color;
-@datepicker-now-background-color: @brand-color-light;
-
-@datepicker-highlight-color: @brand-color-light;
-@datepicker-hover-highlight-color: rgba(0, 0, 0, .04); // 此处亮色暗色通用，无token
-@datepicker-presets-min-width: 100px;
-@datepicker-presets-text-color: @brand-color;
-@datepicker-presets-text-color-hover: @brand-color-hover;
-
-// ani
-@datepicker-transition: box-shadow @anim-duration-base @anim-time-fn-easing, background-color @anim-duration-base linear, border-color @anim-duration-base linear;
