--- style/web/components/date-picker-new/index.html
@@ -1,96 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>组件名称</title>
-
-    <!-- 组件样式（全部的） -->
-    <link rel="stylesheet" href="../../dist/tdesign.css">
-    <!--
-      TDesign 图标库，桌面端与移动端用同样的资源
-      使用方式：<i class="t-icon t-icon-delete"></i>
-    -->
-    <link rel="stylesheet" href="https://tdesign.gtimg.com/icon/0.0.3/fonts/index.css">
-
-</head>
-<body>
-  <div class="tdesign-demo-wrap">
-
-    <!-- 开发者信息 -->
-    <div class="tdesign-demo-wrap__name">
-      <h1 class="">Button</h1>
-      <p class="tdesign-demo-wrap__info">开发者：</p>
-      <p class="tdesign-demo-wrap__info">创建日期：</p>
-      <p class="tdesign-demo-wrap__info">说明：</p>
-    </div>
-
-    <!-- 组件开始区 -->
-
-
-    <!-- 默认样式 -->
-    <div class="tdesign-demo-item">
-      <h2 class="tdesign-demo-item__title">默认</h2>
-
-      <!-- 灰底区域 可多个 -->
-      <div class="tdesign-demo-item__body">
-
-        <!-- 内容区 Start-->
-
-        <!-- 内容区 End-->
-
-      </div>
-    </div>
-
-    <!-- 组件状态 -->
-    <div class="tdesign-demo-item">
-      <h2 class="tdesign-demo-item__title">状态</h2>
-
-      <!-- 灰底区域 可多个 -->
-      <div class="tdesign-demo-item__body">
-
-        <!-- 内容区 Start-->
-
-        <!-- tdesign-demo-block block内容容器，加了上间距 -->
-        <div class="tdesign-demo-block">
-
-        </div>
-
-        <!-- 内容区 End-->
-
-      </div>
-    </div>
-
-    <!-- 组件类型 -->
-    <div class="tdesign-demo-item">
-      <h2 class="tdesign-demo-item__title">类型</h2>
-      <div class="tdesign-demo-item__body">
-
-        <!-- 内容区 Start-->
-
-        <!-- 内容区 End-->
-
-      </div>
-      <div class="tdesign-demo-item__body">
-
-        <!-- 内容区 Start-->
-
-        <!-- 内容区 End-->
-
-      </div>
-    </div>
-
-    <!-- 组件尺寸 -->
-    <div class="tdesign-demo-item">
-      <h2 class="tdesign-demo-item__title">尺寸(如果有)</h2>
-      <div class="tdesign-demo-item__body">
-
-        <!-- 内容区 Start-->
-
-        <!-- 内容区 End-->
-    </div>
-
-  </div>
-
-</body>
-</html>
