--- style/web/components/date-picker/_docs.less
@@ -1,6 +1,0 @@
-.@{d-prefix}-demo-item--date-picker {
-  .@{prefix}-date-picker,
-  .@{prefix}-date-range-picker {
-    max-width: 320px;
-  }
-}
